To fetch all user data from backend database
To fetch all user data such as name, email, bio, image, phoneNumber etc. Because initially we were only getting name, email, sub and some other info when we debugged that in console but we didn't receive all info that we have defined in 'user' SQL table in Supabase. So we fixed that problem with this and some other code as well, via an API call ( also refer : .\app\_layout.jsx )